+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # adrez-mapper Revisions
 
+2020-04-29
+- better streetNumber parser
+
 2020-04-18 (0.6)
 - changed LocatorContact and LocatorCampaign so empty field will remain in the locator.
 - LocatorContact / Camptaign field are only removed if undefined